Dropped ProxyPass support
