+++ html/startPage/index.html
@@ -20,7 +20,6 @@
             <select id="platform" onchange="platformChange()">
                 <option value="java">Java Edition</option>
                 <option value="bedrock">Bedrock Edition</option>
-                <option value="bedrock-proxypass">Bedrock Edition (ProxyPass)</option>
             </select>
         </li>
         <li class="form-row">
@@ -113,9 +112,6 @@
                 <option value="1.16.220">1.16.220</option>
                 <option value="1.16.210">1.16.210</option>
                 <option value="1.16.201">1.16.201</option>
-            </select>
-            <select id="version-bedrock-proxypass" style="display: none;">
-                <option value="1.15.2">1.16.221</option>
             </select>
         </li>
         <li class="form-row">
