+++ src/proxy/bedrock/proxy.js
@@ -1,41 +1,19 @@
-const { spawn } = require('child_process')
-const WebSocket = require('ws')
-const net = require('net')
-
-// const java = require('java')
-
-/* java.asyncOptions = {
-  asyncSuffix: undefined,
-  syncSuffix: "",
-  promiseSuffix: "Promise",
-  promisify: require('util').promisify
-} */
-
-let child
-
-// TODO: Can it still be frozen?
-// let mayBeFrozen = false
-// let timeFrozen
-
-let proxyPass
-let proxyPlayerSession
-
-let wsPort
-let ws
+const { Relay } = require('bedrock-protocol')
 
 let scriptingEnabled = false
 
-// This whole thing is messy for now.
+// http://prismarinejs.github.io/minecraft-data/?v=bedrock_1.19.62&d=protocol#packet_server_to_client_handshake
+// http://prismarinejs.github.io/minecraft-data/?v=bedrock_[VERSION]&d=protocol#[PACKET]
 
 exports.capabilities = {
   modifyPackets: true,
   jsonData: true,
-  rawData: true,
+  rawData: false,
   scriptingSupport: false,
   clientboundPackets: {},
   serverboundPackets: {},
   wikiVgPage: 'https://wiki.vg/Bedrock_Protocol',
-  versionId: 'bedrock-proxypass-json'
+  versionId: undefined
 }
 
 let host
@@ -46,144 +24,8 @@ let messageCallback
 let dataFolder
 let updateFilteringCallback
 
-// https://stackoverflow.com/questions/28050171/nodejs-random-free-tcp-ports
-function freePort () {
-  return new Promise((resolve, reject) => {
-    const srv = net.createServer(function(sock) {
-      sock.end()
-    })
-    srv.listen(0, function() {
-      const port = srv.address().port
-      srv.close()
-      resolve(port)
-    })
-    srv.on('error', (err) => {
-      reject(err)
-    })
-  })
-}
-
-async function launch() {
-  wsPort = Number(await freePort())
-
-  child = spawn('java', ['-jar', dataFolder + '/proxypass/' + 'proxypass-pakkit.jar', '--start-from-args', '0.0.0.0',
-    listenPort.toString(), host, port.toString(), '1', 'true', 'true', 'pakkit', 'pakkitProxyPoweredByProxyPass',
-    'true', wsPort.toString()])
-
-  child.stdout.on('data', handleOutput)
-  child.stderr.on('data', handleError)
-  child.on('close', (code, signal) => {
-    setTimeout(() => {
-      launch()
-    }, 50)
-  })
-
-
-  // setTimeout(startWebsocket, 3000)
-}
-
-function startWebsocket () {
-  ws = new WebSocket('ws://localhost:' + wsPort)
-
-  /* ws.on('open', function open() {
-    ws.send('something')
-  }); */
-
-  ws.on('message', function incoming(data) {
-    // console.log(data);
-    try {
-      data = JSON.parse(data)
-    } catch (err) {
-      console.error(err)
-      return
-    }
-    switch (data.type) {
-      case 'packet':
-        handlePacket(data.data)
-        break
-      case 'event':
-        handleEvent(data)
-        break
-      default:
-        console.log('Unknown message type', data.type)
-    }
-  })
-
-  console.log('Proxy started (Bedrock)!')
-}
-
-function handlePacket (packet) {
-  const name = packet.packetType.toLowerCase()
-
-  const data = JSON.parse(packet.jsonData);
-  const hexIdString = '0x' + packet.packetId.toString(16).padStart(2, '0')
-
-  // These values are unneeded or are exposed elsewhere in the GUI
-  delete data.packetId
-  delete data.packetType
-  delete data.clientId
-  delete data.senderId
-
-  const raw = Object.values(packet.bytes)
-  // Prepend packet ID for consistency with Java Edition
-  raw.unshift(packet.packetId)
-
-  // If the packet as already handled bya  custom handler then scripting cannot modify it
-  // (well it can but that would be annoying to add)
-  const canUseScripting = !data.isHandled
-
-  // TODO: check validity
-  packetCallback(packet.direction, { name: name, className: packet.className }, data, hexIdString, raw, canUseScripting, true)
-}
-
-function handleEvent (event) {
-  switch(event.eventType) {
-    case 'unableToConnect':
-      messageCallback('Unable to connect to server', 'Unable to connect to the Bedrock server at ' +
-        event.eventData.replace(/^\//, '') + // Remove slash at start
-        '. Make sure the server is online.')
-      relaunch()
-      break;
-    case 'disconnect':
-      console.log('Disconnect - relaunching proxy')
-      relaunch()
-      break;
-    case 'filteringPackets':
-      console.log('rec')
-      const packetTypes = JSON.parse(event.eventData)
-      for (const index in packetTypes) {
-        const idString = '0x' + Number(index).toString(16).padStart(2, '0')
-        const name = packetTypes[index].toLowerCase()
-        // There isn't much of a distinction between serverbound and clientbound in Bedrock and many packets can be sent both ways
-        exports.capabilities.clientboundPackets[idString] = name
-        exports.capabilities.serverboundPackets[idString] = name
-      }
-      updateFilteringCallback()
-      break;
-    default:
-      console.log('Unknown event', event.eventType)
-  }
-}
-
-function handleOutput (chunk) {
-  try {
-    const text = chunk.toString('utf8').trim()
-    console.log('ProxyPass output:', text)
-    if (text.startsWith('ProxyPass - Websocket started on port: ') && !(ws && ws.readyState === WebSocket.OPEN)) {
-      setTimeout(startWebsocket, 100)
-    }
-  } catch (err) {
-    console.error(err)
-  }
-}
-
-function handleError (chunk) {
-  try {
-    console.log('ProxyPass error:', chunk.toString('utf8').trim())
-  } catch (err) {
-    console.error(err)
-  }
-}
+let relay
+let relayPlayer
 
 exports.startProxy = function (passedHost, passedPort, passedListenPort, version, onlineMode, authConsent, passedPacketCallback,
                                passedMessageCallback, passedDataFolder, passedUpdateFilteringCallback, authCodeCallback) {
@@ -195,38 +37,75 @@ exports.startProxy = function (passedHost, passedPort, passedListenPort, version
   dataFolder = passedDataFolder
   updateFilteringCallback = passedUpdateFilteringCallback
 
-  launch()
+  // TODO: add dynamic version loading as in bedrock-packet-interceptor
+
+  const mcdata = require('minecraft-data')('bedrock_' + version) // Used to get packets, may remove if I find a better way
+
+  toClientMappings = mcdata.protocol.types.mcpe_packet[1][0].type[1].mappings
+  toServerMappings = mcdata.protocol.types.mcpe_packet[1][0].type[1].mappings
+
+  exports.capabilities.clientboundPackets = toClientMappings
+  exports.capabilities.serverboundPackets = toServerMappings
+
+  // TODO: minecraft-data docs should replace wikiVgPage
+  //exports.capabilities.wikiVgPage = "http://prismarinejs.github.io/minecraft-data/?v=bedrock_{VERSION}&d=protocol".replace("{VERSION}", version)
+
+  exports.capabilities.versionId = 'node-bedrock-protocol-' + version.split('.').join('-')
+
+  relay = new Relay({
+    version: version, // The version
+    /* host and port to listen for clients on */
+    host: '0.0.0.0',
+    port: Number(listenPort),
+    //offline: true,
+    offline: !onlineMode,
+    /* Where to send upstream packets to */
+    destination: {
+      host: host,
+      port: Number(port)
+    },
+    // TODO: use minecraftFolder ?
+    profilesFolder: authConsent ? './profiles/bedrock/default' : dataFolder,
+    onMsaCode: function (data) {
+      console.log('MSA code:', data.user_code)
+      authCodeCallback(data)
+    }
+  })
+
+  relay.conLog = console.debug
+  relay.listen() // Tell the server to start listening.
+
+  relay.on('connect', player => {
+    relayPlayer = player
+    console.log('New connection', player.connection.address)
+  
+    // Server is sending a message to the client.
+    player.on('clientbound', ({ name, params }) => {
+      // TODO: check validity
+      packetCallback('clientbound', { name: name }, params, '0x00', {}, scriptingEnabled, true)
+    })
+    // Client is sending a message to the server
+    player.on('serverbound', ({ name, params }) => {
+      packetCallback('serverbound', { name: name }, params, '0x00', {}, scriptingEnabled, true)
+    })
+  })
 }
 
 exports.end = function () {
-  child.kill()
-}
-
-// used to relaunch on disconnect
-function relaunch () {
-  ws.close()
-  // Will auto-restart
-  child.kill()
+  // TODO
+  relay.close()
 }
 
 exports.writeToClient = function (meta, data) {
-  ws.send(JSON.stringify({
-    type: 'inject',
-    className: meta.className,
-    direction: 'client',
-    data: data
-  }))
-  // proxyPlayerSession.injectPacketStaticPromise(JSON.stringify(data), meta.className, 'client')
+  if (relayPlayer) {
+    relayPlayer.queue(meta.name, data)
+  }
 }
 
 exports.writeToServer = function (meta, data) {
-  // proxyPlayerSession.injectPacketStaticPromise(JSON.stringify(data), meta.className, 'server')
-  ws.send(JSON.stringify({
-    type: 'inject',
-    className: meta.className,
-    direction: 'server',
-    data: data
-  }))
+  if (relayPlayer) {
+    relayPlayer.upstream.queue(meta.name, data)
+  }
 }
 
 // TODO
@@ -234,3 +113,6 @@ exports.writeToServer = function (meta, data) {
   scriptingEnabled = isEnabled
   proxyPlayerSession.setDontSendPacketsPromise(scriptingEnabled)
 } */
+exports.setScriptingEnabled = function (isEnabled) {
+  scriptingEnabled = isEnabled
+}
